refactor: layering
